rename: `AxisClipSliders` -> `BottomPanelSliders`
