+++ src/aics-image-viewer/components/BottomPanelSliders/index.tsx
