+++ src/aics-image-viewer/components/BottomPanelSliders/styles.css
